[SHOP-12] edit main pages
