--- pages/_app.tsx
@@ -1,5 +1,0 @@
-function MyApp({ Component, pageProps }) {
-  return <Component ponent {...pageProps} />;
-}
-
-export default MyApp;
