+++ src/pages/_app.tsx
@@ -1,0 +1,7 @@
+import type { AppProps } from 'next/app';
+
+function MyApp({ Component, pageProps }: AppProps) {
+  return <Component ponent {...pageProps} />;
+}
+
+export default MyApp;
